+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release notes
 
+## New in git-machete 3.19.0
+
+- improved: `git machete github create-pr` also checks for `.github/pull_request_template.md` for description (contributed by @raylu)
+
 ## New in git-machete 3.18.3
 
 - added: arm64 packages in Ubuntu PPA